sdhc: add ipc call IPC_SDHC_EXIT
In some cases it is desirable to let the Broadway processor take complete
control of hardware initially managed by 'mini'.
Add a new IPC call to instruct 'mini' to relinquish control of the
SD Host Controller.

One immediate user of this IPC call is Linux on the Nintendo Wii which
takes complete control of the SDHC hardware.

Signed-off-by: Albert Herranz <albert_herranz@yahoo.es>